Added solution - LeetHub

+++ 17. Rearrange an array with O(1) extra space - GFG/17.-rearrange-an-array-with-o1-extra-space.java
@@ -48,7 +48,7 @@ class Solution
     static void arrange(long arr[], int n)
     {for(int i=0;i<n;i++)
         {
-            arr[i]=arr[i]+(arr[(int)arr[(int)i]]%n)*n;
+            arr[i]=arr[i]+(arr[(int)arr[i]]%n)*n;
            
         }
         for(int i=0;i<n;i++)
